+++ src/instruction/swoc1.cpp
@@ -12,8 +12,7 @@ Simulator::State Simulator::swoc1(Instruction inst)
     new_state.pc += 4;
     try {
         auto pre_mem = m_memory.at(addr);
-        m_memory.at(addr)
-            = static_cast<int32_t>(ftob(m_state_iter->freg.at(op.rs)));
+        m_memory.at(addr) = ftob(m_state_iter->freg.at(op.rs));
         new_state.memory_patch = {true, addr, pre_mem};
     } catch (std::out_of_range e) {
         FAIL("# Memory index out of range\n" << e.what());
